Explicit type conversion (II)

+++ 03-types/main.go
@@ -5,6 +5,6 @@ import "fmt"
 func main() {
 	i := 3                // int
 	j := 2.3              // float
-	sum := float32(i) + j // ERROR: infered type for j is 'float64'!!!
-	fmt.Println(sum)
+	sum := float64(i) + j // Explicit type conversion
+	fmt.Println(sum)      // prints 5.3
 }
